Relationships

+++ database/migrations/2023_07_24_051928_create_replies_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('replies', function (Blueprint $table) {
             $table->id();
             $table->text('body');
-            $table->integer('question_id')->unsigned();
-            $table->integer('user_id');
+            $table->bigInteger('question_id')->unsigned();
+            $table->bigInteger('user_id');
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
             $table->timestamps();
         });
